Debug NVM

+++ src/nvm.c
@@ -30,7 +30,7 @@ enum nvm_memory_status
 
 #define NVM_EXTRACT_MEM_STS(val)  ((uint8_t)(((val) >> 60) & 0x0F))
 #define NVM_IS_WRITTEN(val)       (NVM_EXTRACT_MEM_STS(val) == NVM_MEMORY_WRITTEN)
-#define NVM_WRITE_MEM_STS(val)    (((uint64_t)val & 0x0FFFFFFFFFFFFFFF) | (NVM_MEMORY_WRITTEN << 60))
+#define NVM_WRITE_MEM_STS(val)    ((((uint64_t)val) & 0x0FFFFFFFFFFFFFFF) | (((uint64_t)NVM_MEMORY_WRITTEN) << 60))
 
 // Private variables
 static uint64_t nvm_serial_number_raw;
